+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "julia.executablePath": "julia +1.9.4"
+    "julia.executablePath": "julia"
 }
