+++ Project.toml
@@ -1,3 +1,8 @@
+name = "GeoCombSurfX"
+uuid = "78b1d9a6-0924-4921-b6e7-d9c827ccfda8"
+version = "0.1.0"
+authors = ["Sascha Stüttgen <sascha.stuettgen@rwth-aachen.de", "Reymond Akpanya <reymond.akpanya@art.rwth-aachen.de>"]
+
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 GenericLinearAlgebra = "14197337-ba66-59df-a3e3-ca00e7dcff7a"
@@ -10,6 +15,9 @@ LanguageServer = "2b0e0bc5-e4fd-59b4-8912-456d1b03d8d7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-PolygonOps = "647866c9-e3ac-4575-94e7-e3d426903924"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
+Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+
+[compat]
+julia = "1.10"
