+++ src/affine_geometry.jl
@@ -1,4 +1,20 @@
-using LinearAlgebra
+struct Ray{T<:Real}
+    point::Vector{T}
+    vector::Vector{T}
+
+    function Ray(point::AbstractVector{<:Real}, vector::AbstractVector{<:Real})
+        return new{Float64}(point, vector)
+    end
+end
+
+struct Plane{T<:Real}
+    point::Vector{T} 
+    vectors::Matrix{T}
+
+    function Plane(point::AbstractVector{<:Real}, vectors::AbstractMatrix{<:Real})
+        return new{Float64}(point, vectors)
+    end
+end
 
 # """
 # A::Matrix
@@ -184,24 +200,6 @@ TBW
 """
 function rigidmap(preim::Vector{<:Vector{<:Real}}, im::Vector{<:Vector{<:Real}}; atol::Real = 1e-8)
     return rigidmap(hcat(preim...), hcat(im...), atol = atol)
-end
-
-struct Ray{T<:Real}
-    point::Vector{T}
-    vector::Vector{T}
-
-    function Ray(point::AbstractVector{<:Real}, vector::AbstractVector{<:Real})
-        return new{Float64}(point, vector)
-    end
-end
-
-struct Plane{T<:Real}
-    point::Vector{T} 
-    vectors::Matrix{T}
-
-    function Plane(point::AbstractVector{<:Real}, vectors::AbstractMatrix{<:Real})
-        return new{Float64}(point, vectors)
-    end
 end
 
 function Plane(points::AbstractMatrix{<:Real}; atol::Real=1e-8)
